Processor/requirements.txt
MIDIUtil>=1.2.1
Pillow>=9.1.0
